move mapToObject to src/core/utils
