+++ src/core/utils/index.ts
@@ -18,6 +18,7 @@
  */
 
 export * from './get';
+export * from './map_to_object';
 export * from './pick';
 export * from './assert_never';
 export * from './url';
