+++ src/core/utils/map_to_object.ts
